+++ nano-vllm-java/src/main/java/com/igormaznitsa/nanollvm/tensor/tornado/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Optional TornadoVM GEMV offload for {@link com.igormaznitsa.nanollvm.tensor.FloatKernels}.
+ *
+ * @since 1.3.1
+ */
+
+package com.igormaznitsa.nanollvm.tensor.tornado;